requirements.txt
Flask==1.1.2
Flask-Login==0.5.0
Flask-Webpack==0.1.0
mediacloud==3.11.3
pymongo==3.10.1
requests==2.23.0
raven[flask]==6.10.0
redis==2.10.6
networkx==1.9
Flask-Mail==0.9.1
mediacloud-cliff==2.6.1
gunicorn==20.0.4
dogpile.cache==0.7.1  # upgrading dogpile.cache caused problems - leave it on this version
python-dateutil==2.8.1
python-slugify==4.0.0
greenlet==0.4.15
deco==0.5.2
Flask-Executor==0.9.3
newspaper3k==0.2.8
pytest==5.4.1
gevent==20.5.0
